add query synchronization feature
- added synchronize option to config;
- added synchronize option to query meta
- added function for manual cache update

Queries now sync data unless the synchronize option is set to false in the config or query meta.
Since data propagation occurs by checking the "manual" property in the event action properties, manually setting the cache via the QueryClient will not propagate the data change to other queries. In order to manually setting the cache and propagate the changes, the function of setting the cache has been added.